Divide intermediate points by 3
In order to get the beginning and ending velocities of V to match vPre and vTarg, the intermediate points have to be shortened by a factor of 3 for some reason to get everything to match up. Not really sure what is going on here and it is worth exploring more. One thought is that the time is not really 0-1 sec. This visual could be throwing things off.